Rest.Client.Internal

Documentation

module Control.Monad

module Data.String

class Monad m => MonadIO m where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

Methods

liftIO :: IO a -> m a

Lift a computation from the IO monad.

data ByteString :: *

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A lazy ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Lazy.Char8 it can be interpreted as containing 8-bit characters.

intercalate :: [a] -> [[a]] -> [a]

intercalate xs xss is equivalent to (concat (intersperse xs xss)). It inserts the list xs in between the lists in xss and concatenates the result.

encode :: String -> String

module Rest.Client.Base

class ShowUrl a where

Methods

showUrl :: a -> String

hAccept :: HeaderName

hContentType :: HeaderName

data ApiRequest

Constructors

ApiRequest

Fields

method :: String

uri :: String

params :: [(String, String)]

requestHeaders :: RequestHeaders

requestBody :: ByteString

doRequest :: ApiStateC m => (ByteString -> Reason e) -> (ByteString -> a) -> ApiRequest -> m (ApiResponse e a)

parseResult :: (ByteString -> Reason e) -> (ByteString -> a) -> Response ByteString -> ApiResponse e a

fromJSON :: FromJSON a => ByteString -> a

toJSON :: ToJSON a => a -> ByteString

fromXML :: XmlStringToType a => ByteString -> a

toXML :: XmlStringToType a => a -> ByteString

makeReq :: String -> String -> [[String]] -> [(String, String)] -> RequestHeaders -> ByteString -> ApiRequest

doReq :: (ApiStateC m, MonadIO m) => ApiRequest -> m (Response ByteString)